dcos-oauth: fix verified check
The verified check used to fail only if the input was a bool and false.
This fixes the condition so the check also fails if the
response is not a bool at all.

The external component's API would have to change to make this check fail
but since it is there, it might as well be correct.